Backend/requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.2.1
django-heroku==0.3.1
django-restframework==0.0.1
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.5
PyJWT==1.7.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
sqlparse==0.3.1
whitenoise==5.0.1
